requirements.txt
diffusers==0.30.0
pillow==10.4.0
boto3
numpy